+++ src/django_countdown/apps.py
@@ -1,7 +1,8 @@
 from django.apps import AppConfig
+from django.utils.translation import gettext_lazy as _
 
 
 class DjangoCountdownConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "django_countdown"
-    verbose_name = "Odliczanie do zamknięcia serwisu"
+    verbose_name = _("Site shutdown countdown")